+++ ICD.Connect.Settings/Originators/OriginatorConsole.cs
@@ -1,9 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using ICD.Common.Utils;
 using ICD.Connect.API.Commands;
 using ICD.Connect.API.Nodes;
-using ICD.Connect.Telemetry;
 
 namespace ICD.Connect.Settings.Originators
 {