+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-### Added
+### Added 
+ - Failing more gracefully when a duplicate settings factory name is cached.
  - SPlus Shim Request Resync with API
  - SPlusSafeString for safely sending strings to S+
 
@@ -48,6 +49,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
  - Moving originators and settings classes into subdirectories
  - Incremented config version from 3.0 to 3.1
+
+## [7.5.2] - 2019-07-31
+### Changed
+ - Small performance improvement in plugin discovery
+ - Telemetry service is optional
 
 ## [7.5.1] - 2019-06-24
 ### Changed